backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
rapidocr-onnxruntime>=1.3.0
sqlalchemy>=2.0.0
pillow>=10.0.0
numpy>=1.26.0
python-multipart>=0.0.9
